fixed routes logic
